requirements.txt
Flask
Flask-JWT-Extended
Flask-RESTful
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Werkzeug
flask-restx